requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2

# Azure Services
azure-core>=1.29.4
azure-identity>=1.14.1
azure-storage-file-datalake>=12.12.0
azure-storage-blob>=12.18.3
azure-search-documents==11.6.0b8
azure-ai-textanalytics>=5.3.0
azure-ai-vision-imageanalysis==1.0.0
azure-cognitiveservices-language-textanalytics>=0.2.0

# OpenAI
openai>=1.3.0

# Document & Data Processing
PyMuPDF>=1.23.4
Pillow>=10.0.1
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.1.1
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
ffmpeg-python>=0.2.0

# Testing (optional)
pytest>=7.4.2
pytest-asyncio>=0.21.1
beautifulsoup4>=4.11.0
lxml>=4.9.3

langdetect